delete all

+++ stankinsv2/solution/StankinsV2/StankinsData/Writables.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace StankinsDataWeb
@@ -22,18 +23,37 @@ namespace StankinsDataWeb
         }
         public async Task<string> GetFileContents(string nameFile)
         {
+            if(!Verify(nameFile)){
+                return null;
+            }
+
             nameFile=Path.Combine(folder,nameFile);
             return await File.ReadAllTextAsync(nameFile);
         }
         public async Task WriteFileContents(string nameFile,string content)
         {
+            if(!Verify(nameFile)){
+                return ;
+            }
             nameFile=Path.Combine(folder,nameFile);
             await File.WriteAllTextAsync(nameFile,content);
         }
         public void DeleteFile(string nameFile)
         {
+            if(!Verify(nameFile)){
+                return ;
+            }
             nameFile=Path.Combine(folder,nameFile);
             File.Delete(nameFile);
+        }
+        private static bool Verify(string id)
+        {
+            // Restrict the username to letters and digits only
+            if (!Regex.IsMatch(id, "^[a-zA-Z0-9]+$"))
+            {
+                return false;
+            }
+            return true;
         }
     }
 }
